refactor script
