+++ packages/jellyfish-transaction/src/script/mapping.ts
@@ -1,4 +1,7 @@
-import { StaticCode } from './opcode'
+import { SmartBuffer } from 'smart-buffer'
+import { readVarUInt, writeVarUInt } from '../buffer/buffer_varuint'
+import { toBuffer, toOPCodes } from './_buffer'
+import { OPCode, StaticCode } from './opcode'
 import { OP_PUSHDATA } from './data'
 import { OP_DEFI_TX } from './defi'
 import { CDfTx, DfTx } from './defi/dftx'
@@ -57,6 +60,61 @@ export class OP_UNMAPPED extends StaticCode {
  * @see https://github.com/DeFiCh/ain/blob/master/src/script/script.h
  */
 export const OP_CODES = {
+  /**
+   * Read SmartBuffer and create OPCode[] stack.
+   *
+   * Using P2WPKH redeem script as an example.
+   *
+   * Input Example: 1600140e7c0ab18b305bc987a266dc06de26fcfab4b56a
+   *   0x16 (VarUInt)
+   *   0x00 (OP_0)
+   *   6ab5b4fafc26de06dc66a287c95b308bb10a7c0e (formatted as big endian)
+   *
+   * Output Example:
+   *   OP_0
+   *   OP_PUSHDATA<RIPEMD160(SHA256(pubkey))>
+   *
+   * @param buffer {SmartBuffer} to read from
+   * @return {OPCode[]} read from buffer to OPCode
+   */
+  fromBuffer (buffer: SmartBuffer) {
+    const length = readVarUInt(buffer)
+    if (length === 0) {
+      return []
+    }
+
+    return toOPCodes(SmartBuffer.fromBuffer(buffer.readBuffer(length)))
+  },
+  /**
+   * Converts OPCode[] and write it into SmartBuffer.
+   *
+   * Using P2PKH redeem script as an example.
+   *
+   * Input Example:
+   *   OP_DUP
+   *   OP_HASH160
+   *   OP_PUSHDATA<RIPEMD160(SHA256(pubkey))>
+   *   OP_EQUALVERIFY
+   *   OP_CHECKSIG
+   *
+   * Output Example: 1976a9143bde42dbee7e4dbe6a21b2d50ce2f0167faa815988ac
+   *   0x19 (VarUInt)
+   *   0x76 (OP_DUP)
+   *   0xa9 (OP_HASH160)
+   *   5981aa7f16f0e20cd5b2216abe4d7eeedb42de3b (formatted as big endian)
+   *   0x88 (OP_EQUALVERIFY)
+   *   0xac (OP_CHECKSIG)
+   *
+   * @param stack {OPCode[]} to convert into raw buffer
+   * @param buffer {SmartBuffer} to write to
+   */
+  toBuffer (stack: OPCode[], buffer: SmartBuffer) {
+    const buffs = toBuffer(stack)
+
+    // Write the len of buffer in bytes and then all the buffer
+    writeVarUInt(buffs.length, buffer)
+    buffer.writeBuffer(buffs)
+  },
   OP_DEFI_TX: (dftx: DfTx<any>): OP_DEFI_TX => {
     return new OP_DEFI_TX(dftx)
   },
